// test-crm/src/modules/profile/profile.interface.ts
export interface ClientProfileData {
  id: any;
  userProfileId: string;
  userProfileVersion: any;
  firstName: string;
  lastName: string;
  email: string;
  document: string;
  corporateDocument: any;
  corporateName: any;
  documentType: string;
  phone: string;
  tradeName: any;
  stateInscription: string;
  corporatePhone: any;
  isCorporate: string;
  customerClass: any;
  customerCode: any;
}